chore: moved the files to config folder to align int_adyen_SFRA
