+++ src/cartridges/app_adyen_SFRA/cartridge/client/default/js/adyen_checkout/paymentMethodsConfiguration/giftcards/giftcardsConfig.js
@@ -10,7 +10,7 @@ const {
   GIFTCARD,
   SUCCESS,
   NOTENOUGHBALANCE,
-} = require('../../../../../../utils/constants');
+} = require('../../../../../../config/constants');
 const { initializeCheckout } = require('../../renderGenericComponent');
 
 class GiftCardConfig {
