+++ src/cartridges/app_adyen_SFRA/cartridge/client/default/js/adyen_checkout/paymentMethodsConfiguration/index.js
@@ -10,7 +10,7 @@ const ApplePayConfig = require('./applePay/applePayConfig');
 const AmazonPayConfig = require('./amazonPay/amazonPayConfig');
 const PayPalConfig = require('./paypal/paypalConfig');
 const GiftCardsConfig = require('./giftcards/giftcardsConfig');
-const store = require('../../../../../utils/store');
+const store = require('../../../../../config/store');
 const helpers = require('../helpers');
 const { httpClient } = require('../../commons/httpClient');
