+++ src/cartridges/app_adyen_SFRA/cartridge/client/default/js/adyen_express/cart/expressPayments.js
@@ -9,7 +9,7 @@ const {
   GOOGLE_PAY,
   PAYPAL,
   AMAZON_PAY,
-} = require('../../../../../utils/constants');
+} = require('../../../../../config/constants');
 
 function getPaymentMethodConfig(adyenPaymentMethods, paymentMethodType) {
   return adyenPaymentMethods?.paymentMethods.find(
