+++ src/cartridges/app_adyen_SFRA/cartridge/client/default/js/adyen_express/googlePayExpress.js
@@ -8,7 +8,7 @@ const {
   GOOGLE_PAY,
   PAY_WITH_GOOGLE,
   GOOGLE_PAY_CALLBACK_TRIGGERS,
-} = require('../../../../utils/constants');
+} = require('../../../../config/constants');
 const { httpClient } = require('../commons/httpClient');
 const { initializeCheckout } = require('./initializeCheckout');
