+++ src/cartridges/app_adyen_SFRA/cartridge/client/default/js/adyen_express/paymentMethods/amazonpay/amazonPayExpressPart2.js
@@ -1,5 +1,5 @@
 const { getPaymentMethods } = require('../../../commons');
-const { AMAZON_PAY } = require('../../../../../../utils/constants');
+const { AMAZON_PAY } = require('../../../../../../config/constants');
 const { httpClient } = require('../../../commons/httpClient');
 
 async function saveShopperDetails(details) {
